Default clean-up

+++ Arduino Sketches/Main/Spiderwort_Incubator_V1.9/Incuvers_Settings.h
@@ -16,7 +16,7 @@ struct HardwareStruct {
   byte ident[3];
   // Identification
   byte hVer[3];
-  byte serial;
+  int serial;
   // Temp settings
   byte countOfTempSensors;
   byte sensorAddrDoorTemp[8];
